requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
jinja2
python-dotenv
openai
requests
